pyRPPG/requirements.txt
# for basic numerics
numpy==1.16.4
scipy==1.4.1
Bottleneck==1.3.1
pyEDFlib==0.1.15
PyWavelets==1.1.1
matplotlib==3.1.2

# for machine learning and computer vision
tensorflow-gpu==1.14.0
opencv-python==4.1.1.26
dlib==19.19.0


# for download models
wget==3.2
gdown==3.12.2


